174 moved module to guice package; missing license headers added
